Extract frontend tests into separate file
